+++ python_modules/dagster/dagster/core/definitions/config.py
@@ -3,7 +3,6 @@ from collections import namedtuple
 from dagster import check
 from dagster.builtins import BuiltinEnum
 from dagster.config.field_utils import check_user_facing_opt_config_param
-from dagster.core.execution.config import IRunConfig
 from dagster.primitive_mapping import is_supported_config_python_builtin
 
 
@@ -16,14 +15,14 @@ class ConfigMapping(namedtuple('_ConfigMapping', 'config_fn config_field')):
 
     By specifying a config mapping function, you can override the configuration for the child
     solids contained within a composite solid.
-            
+
     Config mappings require the configuration field to be specified as ``config``, which will be
     exposed as the configuration field for the composite solid, as well as a configuration mapping
     function, ``config_fn``, which maps the config provided to the composite solid to the config
     that will be provided to the child solids.
-    
+
     Args:
-        config_fn (Callable[[ConfigMappingContext, dict], dict]): The function that will be called
+        config_fn (Callable[[dict], dict]): The function that will be called
             to map the composite config to a config appropriate for the child solids.
         config_field (Field): The schema of the composite config.
     '''
@@ -35,17 +34,4 @@ class ConfigMapping(namedtuple('_ConfigMapping', 'config_fn config_field')):
             cls,
             config_fn=check.callable_param(config_fn, 'config_fn'),
             config_field=check_user_facing_opt_config_param(config, 'config'),
-        )
-
-
-class ConfigMappingContext(namedtuple('ConfigMappingContext', 'run_config')):
-    '''Config mapping-specific context.
-
-    Attributes:
-        run_config (RunConfig): The run config belonging to this pipeline run.
-    '''
-
-    def __new__(cls, run_config=None):
-        return super(ConfigMappingContext, cls).__new__(
-            cls, check.inst_param(run_config, 'run_config', IRunConfig)
         )