+++ python_modules/dagster/dagster_tests/core_tests/execution_plan_tests/test_execution_plan_composite.py
@@ -27,7 +27,7 @@ def composite_pipeline():
 
 
 @composite_solid(
-    config_fn=lambda _, cfg: {
+    config_fn=lambda cfg: {
         'node_a': {'config': {'foo': cfg['foo']}},
         'node_b': {'config': {'bar': cfg['bar']}},
     },